Split resolver validation tests
